chore: extract the global shared Borsh schema into a separate module
