+++ packages/legacy/library/src/publickey.ts
@@ -2,9 +2,8 @@ import BN from 'bn.js';
 import bs58 from 'bs58';
 import {Buffer} from 'buffer';
 import {sha256} from '@noble/hashes/sha256';
-import {isOnCurve} from 'legacy-shared';
+import {SOLANA_SCHEMA, Struct, isOnCurve} from 'legacy-shared';
 
-import {Struct, SOLANA_SCHEMA} from './utils/borsh-schema';
 import {toBuffer} from './utils/to-buffer';
 
 /**
