+++ packages/legacy/library/src/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './borsh-schema';
+export {Enum, SOLANA_SCHEMA, Struct} from 'legacy-shared';
 export * from './cluster';
 export * from './send-and-confirm-raw-transaction';
 export * from './send-and-confirm-transaction';
