+++ packages/legacy/shared/package.json
@@ -37,6 +37,8 @@
     "dependencies": {
         "@noble/ed25519": "^2.0.0",
         "@noble/hashes": "^1.3.2",
+        "borsh": "^0.7.0",
+        "buffer": "6.0.3",
         "crypto-impl": "workspace:*"
     },
     "devDependencies": {
