+++ packages/legacy/shared/src/index.ts
@@ -1,1 +1,4 @@
 export * from './ed25519';
+export * from './enum';
+export * from './schema';
+export * from './struct';
