+++ packages/legacy/shared/src/schema.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/ban-types, @typescript-eslint/no-explicit-any
+export const SOLANA_SCHEMA = new Map<Function, any>();
